Fix tests for codegeneration
